ui/rsl: trip load graph cleanup
