+++ ui/rsl/src/components/MiniTripLoadGraph.tsx
@@ -4,138 +4,88 @@ export type MiniTripLoadGraphProps = {
   edges: PaxMonEdgeLoadInfo[];
 };
 
+interface SectionGeometry {
+  x: number;
+  width: number;
+  forecastHeight: number;
+  forecastColor: string;
+  capacityHeight: number;
+  expectedHeight: number;
+}
+
+function getSectionColor(capacity: number, maxLoad: number) {
+  if (capacity == 0) {
+    return "#3C414B";
+  } else {
+    const load = maxLoad / capacity;
+    if (load > 2.0) {
+      return "#C50014";
+    } else if (load > 1.2) {
+      return "#EC0016";
+    } else if (load > 1.0) {
+      return "#F39200";
+    } else if (load > 0.8) {
+      return "#FFD800";
+    } else {
+      return "#408335";
+    }
+  }
+}
+
 function MiniTripLoadGraph({ edges }: MiniTripLoadGraphProps): JSX.Element {
   const graphWidth = 1000;
   const graphHeight = 100;
-  const minEdgeLoad = (eli: PaxMonEdgeLoadInfo) => eli.dist.min;
-  const maxEdgeLoad = (eli: PaxMonEdgeLoadInfo) => eli.dist.max;
-  const avgEdgeLoad = (eli: PaxMonEdgeLoadInfo) => eli.dist.q50;
+
+  const maxEdgeLoad = (eli: PaxMonEdgeLoadInfo) => eli.dist.q95;
 
   const maxCapacity = edges.reduce(
     (max, eli) => (eli.capacity ? Math.max(max, eli.capacity) : max),
     0
   );
-  const maxPax = edges.reduce((max, eli) => Math.max(max, maxEdgeLoad(eli)), 0);
-  const maxLoad = Math.max(maxCapacity, maxPax, 1);
+  const totalMaxPax = edges.reduce(
+    (max, eli) => Math.max(max, maxEdgeLoad(eli)),
+    0
+  );
+  const totalMaxLoad = Math.max(maxCapacity, totalMaxPax, 1);
 
-  const sectionWidth = graphWidth / edges.length;
+  const loadHeight = (load: number) => (load / totalMaxLoad) * graphHeight;
 
-  const sectionXStart = (idx: number) => idx * sectionWidth;
-  const loadY = (load: number) => (1 - load / maxLoad) * graphHeight;
-
-  // TODO:
-  // - diff graph - 2 edge sets
-  // - if trip route was changed, show new route and only matching edges
-  // - show +/- pax per section (min/max/avg?)
-  // - section infos (station names) @hover?
+  const sectionDurations = edges.map((e) =>
+    Math.max(300, e.arrival_schedule_time - e.departure_schedule_time)
+  );
+  const totalDuration = sectionDurations.reduce((sum, v) => sum + v);
+  const sectionGeometry = edges.reduce(
+    (a, eli, idx) => {
+      const capacity = eli.capacity;
+      const maxLoad = maxEdgeLoad(eli);
+      const sg = {
+        x: a.x,
+        width: (sectionDurations[idx] / totalDuration) * graphWidth,
+        forecastHeight: loadHeight(maxLoad),
+        forecastColor: getSectionColor(capacity, maxLoad),
+        capacityHeight: loadHeight(capacity),
+        expectedHeight: loadHeight(eli.expected_passengers),
+      };
+      return { x: sg.x + sg.width, sections: [...a.sections, sg] };
+    },
+    { x: 0, sections: [] as SectionGeometry[] }
+  ).sections;
 
   return (
     <svg viewBox={`0 0 ${graphWidth} ${graphHeight}`} className="w-full">
       <g>
-        {edges.map((eli, idx) => {
-          const capacity = eli.capacity;
-          const left = sectionXStart(idx);
-          const bgs = [];
-          if (capacity) {
-            const yRed = 0;
-            const yYellow = loadY(capacity);
-            const yGreen = loadY(capacity * 0.8);
-            const hRed = yYellow - yRed;
-            const hYellow = yGreen - yYellow;
-            const hGreen = graphHeight - yGreen;
-
-            if (hRed > 0) {
-              bgs.push(
-                <rect
-                  x={left}
-                  y={yRed}
-                  width={sectionWidth}
-                  height={hRed}
-                  fill="#FFCACA"
-                  key="red"
-                />
-              );
-            }
-            if (hYellow > 0) {
-              bgs.push(
-                <rect
-                  x={left}
-                  y={yYellow}
-                  width={sectionWidth}
-                  height={hYellow}
-                  fill="#FFF3CA"
-                  key="yellow"
-                />
-              );
-            }
-            if (hGreen > 0) {
-              bgs.push(
-                <rect
-                  x={left}
-                  y={yGreen}
-                  width={sectionWidth}
-                  height={hGreen}
-                  fill="#D4FFCA"
-                  key="green"
-                />
-              );
-            }
-          } else {
-            bgs.push(
-              <rect
-                x={left}
-                y="0"
-                width={sectionWidth}
-                height={graphHeight}
-                fill="white"
-                key="white"
-              />
-            );
-          }
-          return <g key={idx}>{bgs}</g>;
-        })}
-      </g>
-      <g>
-        {edges.map((eli, idx) => (
-          <path
-            key={idx}
-            className="fill-indigo-500 opacity-40"
-            d={`M ${sectionXStart(idx)} ${loadY(
-              maxEdgeLoad(eli)
-            )} h ${sectionWidth} V ${loadY(
-              minEdgeLoad(eli)
-            )} h ${-sectionWidth} Z`}
-          />
+        {sectionGeometry.map((sg, idx) => (
+          <g key={idx}>
+            <rect
+              x={sg.x}
+              y={graphHeight - sg.forecastHeight}
+              width={sg.width}
+              height={sg.forecastHeight}
+              fill={sg.forecastColor}
+            />
+          </g>
         ))}
       </g>
-      <g>
-        {edges.map((eli, idx) => (
-          <path
-            key={idx}
-            className="stroke-indigo-800 stroke-2"
-            d={`M ${sectionXStart(idx)} ${loadY(
-              avgEdgeLoad(eli)
-            )} h ${sectionWidth}`}
-          />
-        ))}
-      </g>
-      <g>
-        {edges.map((_, idx) => (
-          <path
-            key={idx}
-            stroke="#aaa"
-            d={`M ${sectionXStart(idx)} 0 V ${graphHeight}`}
-          />
-        ))}
-      </g>
-      <rect
-        x="0"
-        y="0"
-        width={graphWidth}
-        height={graphHeight}
-        fill="transparent"
-        stroke="#333"
-      />
     </svg>
   );
 }
